requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioredis==2.0.1
aiosignal==1.3.1
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
aspose-words==24.11.0
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.2.0
azure-core==1.30.2
azure-identity==1.17.1
azure-servicebus==7.12.2
azure-storage-blob==12.24.0
bitarray==3.0.0
black==24.8.0
blinker==1.8.2
Brotli==1.1.0
bytecode==0.16.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
clear==2.0.0
click==8.1.7
cohere==5.13.12
colbert-ai==0.2.21
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
ddtrace==2.14.2
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.15
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dnspython==2.6.1
doclayout_yolo==0.0.3
email_validator==2.2.0
envier==0.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.1
fastapi-cli==0.0.5
fastapi-restful==0.6.0
fastapi-utils==0.7.0
fastavro==1.10.0
filelock==3.16.1
flake8==7.1.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.3
fuzzywuzzy==0.18.0
gevent==24.2.1
geventhttpclient==2.3.1
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.0.3
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
identify==2.6.4
idna==3.7
importlib_metadata==8.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.8
locust==2.29.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnxruntime==1.19.2
openai==1.42.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
openpyxl==3.1.2
opentelemetry-api==1.29.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
plum-dispatch==1.7.4
portalocker==2.10.1
pre_commit==4.0.1
prompt_toolkit==3.0.47
propcache==0.3.0
protobuf==5.29.2
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
pyodbc==5.1.0
pyparsing==3.2.1
pypdf==4.3.1
pypdfium2==4.30.0
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==7.4.4
pytest-repeat==0.9.3
pytest-rerunfailures==14.0
python-dateutil==2.9.0
python-dotenv==1.0.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
redis==5.0.8
redisearch==2.0.0
redisvl==0.3.3
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rerankers==0.5.3
rich==13.7.1
rmtest==0.7.0
safetensors==0.5.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==3.4.1
shellingham==1.5.4
simsimd==6.2.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
spire==0.4.2
Spire.Presentation==9.11.2
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.38.6
stringzilla==3.12.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.2.0
typer==0.15.2
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.32.0
uvloop==0.21.0
virtualenv==20.29.2
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
langchain-core
langchain-experimental
langchain-text-splitters
langgraph
langgraph-sdk
langsmith
langchain_openai
pillow
fastapi
uvicorn
ftfy
sentence-transformers
transformers
loguru
azure-ai-documentintelligence
pdf2image
huggingface_hub
azure-storage-blob
azure-identity
Spire.Doc
pytesseract
aspose-words
ray
ipykernel
pymupdf4llm
langchain-pymupdf4llm
langdetect
reportlab
langchain-milvus